server: Made the game playable from the server
Now the server has an HTML page (the website) which adds the ability
to play
the game from the website, no need to download anything.

Also migrated the WS lib to one that compiles to WASM as if not I was
not able
to run the game on the browser.