store/units: Have the units movement rely 100% on the server
For some reason the flickering was related to the server units beeing slower than the client units,
ideally the HashPath checking was doing so the units Path could be just calculated on the client
so it would run more smothely.

For some reason I was not able to find the root of the issue so as of now the units are fully controlled
by the server, meaning that on every TPS the unit is on the position in which the server things it it.

Which is not bad by itself